+++ TagsCloudVisualization/WordsProcessing/BoringWordsFilter.cs
@@ -20,4 +20,6 @@ namespace TagsCloudVisualization.WordsProcessing
                 .Then(boringWords => words.Where(word => !boringWords.Contains(word)));
         }
     }
+
+
 }
